backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
websockets
bcrypt
PyJWT
pydantic[email]
httpx
